+++ src/NBitcoin/Formatters/BlockExplorerFormatter.cs
@@ -2,7 +2,7 @@
 using Newtonsoft.Json;
 using Newtonsoft.Json.Linq;
 
-namespace NBitcoin.RPC
+namespace NBitcoin.Formatters
 {
     internal class BlockExplorerFormatter : RawFormatter
     {
