+++ src/NBitcoin/Formatters/SatoshiFormatter.cs
@@ -5,7 +5,7 @@ using NBitcoin.DataEncoders;
 using Newtonsoft.Json;
 using Newtonsoft.Json.Linq;
 
-namespace NBitcoin.RPC
+namespace NBitcoin.Formatters
 {
     internal class SatoshiFormatter : RawFormatter
     {
