+++ src/Stratis.Bitcoin.Features.RPC.Tests/Controller/RPCControllerTest.cs
@@ -14,7 +14,6 @@ using Microsoft.AspNetCore.Mvc.Infrastructure;
 using Microsoft.Extensions.Primitives;
 using Moq;
 using NBitcoin;
-using NBitcoin.RPC;
 using Newtonsoft.Json;
 using Newtonsoft.Json.Linq;
 using Stratis.Bitcoin.Configuration;
