+++ src/Stratis.Bitcoin.Features.RPC/Controllers/RPCController.cs
@@ -8,7 +8,6 @@ using Microsoft.AspNetCore.Mvc.Controllers;
 using Microsoft.AspNetCore.Mvc.Infrastructure;
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Primitives;
-using NBitcoin.RPC;
 using Stratis.Bitcoin.Controllers;
 using Stratis.Bitcoin.Utilities;
 using Stratis.Bitcoin.Utilities.JsonErrors;
