+++ src/Stratis.Bitcoin.Features.RPC/RPCClientFactory.cs
@@ -1,6 +1,5 @@
 ﻿using System;
 using NBitcoin;
-using NBitcoin.RPC;
 using Stratis.Bitcoin.Utilities;
 
 namespace Stratis.Bitcoin.Features.RPC
