+++ src/Stratis.Bitcoin.IntegrationTests/RPC/RPCTests.cs
@@ -1,7 +1,6 @@
 using System;
 using System.Net;
 using NBitcoin;
-using NBitcoin.RPC;
 using Stratis.Bitcoin.Connection;
 using Stratis.Bitcoin.Features.RPC;
 using Stratis.Bitcoin.Features.Wallet;
